Removes custom parser for ron parser.

+++ oatie/transform-test.sh
@@ -6,7 +6,8 @@ cargo build --release
 set +e
 
 for file in in/*; do
-    # do something on $file
+    test -f "${file}" || continue
+    echo
     echo $file
     cat "$file" | RUST_BACKTRACE=1 ../target/release/oatie-transform > /dev/null
 done
